size to int (4 bytes)

+++ lab/sandbox/crazydb/src/br/eti/fml/crazydb/ByteUtil.java
@@ -12,10 +12,10 @@ public final class ByteUtil {
     }
 
     public static final long TB = 1024L * 1024L * 1024L * 1024L;
-    public static final long GB = 1024L * 1024L * 1024L;
-    public static final long MB = 1024L * 1024L;
-    public static final long KB = 1024L;
-    public static final long B = 1L;
+    public static final int GB = 1024 * 1024 * 1024;
+    public static final int MB = 1024 * 1024;
+    public static final int KB = 1024;
+    public static final int B = 1;
 
     public static byte[] UUID2bytes(UUID uuid) {
         return ByteBuffer.allocate(16).putLong(
